edge-device/requirements.txt
awscrt==0.24.1
awsiotsdk @ file:///Users/jalilinayat-hussain/Downloads/aws-iot-device-sdk-python-v2
boto3==1.38.11
botocore==1.38.11
jmespath==1.0.1
logger==1.4
numpy==2.2.5
opencv-python==4.11.0.86
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
